PDF_API/requirements.txt

fastapi
uvicorn
python-multipart
opencv-python-headless
pillow
numpy
